added unit tests for backend; extracted Tree related logic from index.ts and put in a separate file
