+++ jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+    preset: 'ts-jest',
+    testEnvironment: 'jest-environment-jsdom',
+    testMatch: ['**/__tests__/*.test.js'],
+  };
