+++ testing-files/RebuiltTree.js
@@ -1,0 +1,78 @@
+const rebuiltTree = {
+  root: {
+    children: [ {
+        children: [ {
+            children: [ {
+                    children: [],
+                    tagObj: {
+                        tag: 0,
+                        tagName: 'Function Component'
+                    },
+                    componentName: 'A SSC',
+                    componentProps: null,
+                    componentState: null,
+                    _debugHookTypes: null,
+                    key: null,
+                    renderDurationMS: 0.39999985694885254,
+                    actualStartTime: 2888.99000999046326,
+                    selfBaseDuration: 0.09999990463256836
+                }],
+                tagObj: {
+                    tag: 5,
+                    tagName: 'Host Component'
+                },
+                componentName: '',
+                componentProps: {children: 'Welcome To My Application'},
+                componentState: null,
+                _debugHookTypes: null,
+                key: null,
+                renderDurationMS: 0.5999999046325684,
+                actualStartTime: 2887.99000999046326,
+                selfBaseDuration: 0.09999990463256836
+            }, 
+            {
+                children: [],
+                tagObj: {
+                    tag: 0,
+                    tagName: 'Function Component'
+                },
+                componentName: 'A CSC',
+                componentProps: {},
+                componentState: {baseQueue: null,
+                    baseState: "",
+                    memoizedState: ""},
+                _debugHookTypes: ['useState', 'useState', 'useEffect'],
+                key: null,
+                renderDurationMS: 0.6099999046325684,
+                actualStartTime: 2888.49000999046326,
+                selfBaseDuration: 0.09999990463256836
+            }],
+        tagObj: {
+            tag: 0,
+            tagName: 'Function Component'
+        },
+        componentName: 'App',
+        componentProps: null,
+        componentState: null,
+        _debugHookTypes: null,
+        key: null,
+        renderDurationMS: 0.5999999046325684,
+        actualStartTime: 2887.6999999046326,
+        selfBaseDuration: 0.09999990463256836
+    }],
+    tagObj: {
+        tag: 3,
+        tagName: 'Host Root'
+    },
+    componentName: '',
+    componentProps: null,
+    componentState: null,
+    _debugHookTypes: null,
+    key: null,
+    renderDurationMS: 0.5999999046325684,
+    actualStartTime: 2887.5999999046326,
+    selfBaseDuration: 0.09999990463256836
+  }
+}
+
+module.exports = rebuiltTree;
